+++ .idea/workspace.xml
@@ -5,12 +5,11 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="938e4817-0387-47e8-95b5-07acc00b3c4f" name="Changes" comment="added excel file reader">
-      <change afterPath="$PROJECT_DIR$/.idea/php.xml" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/pp-2023/src/app/components/ExcelFileReader.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/pp-2023/src/app/components/ExcelFileUploader.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/pp-2023/Dockerfile" beforeDir="false" afterPath="$PROJECT_DIR$/pp-2023/Dockerfile" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/pp-2023/package-lock.json" beforeDir="false" afterPath="$PROJECT_DIR$/pp-2023/package-lock.json" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/pp-2023/package.json" beforeDir="false" afterPath="$PROJECT_DIR$/pp-2023/package.json" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/pp-2023/src/app/components/ExcelFileReader.js" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/pp-2023/src/app/kmeans/calculateButton.js" beforeDir="false" afterPath="$PROJECT_DIR$/pp-2023/src/app/kmeans/calculateButton.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/pp-2023/src/app/page.js" beforeDir="false" afterPath="$PROJECT_DIR$/pp-2023/src/app/page.js" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
@@ -55,9 +54,9 @@
   <component name="MarkdownSettingsMigration">
     <option name="stateVersion" value="1" />
   </component>
-  <component name="ProjectColorInfo"><![CDATA[{
-  "associatedIndex": 2
-}]]></component>
+  <component name="ProjectColorInfo">{
+  &quot;associatedIndex&quot;: 2
+}</component>
   <component name="ProjectId" id="2VNgx2NKxrpZRtc6X5VmeRLuF6h" />
   <component name="ProjectLevelVcsManager">
     <ConfirmationsSetting value="2" id="Add" />
@@ -66,26 +65,26 @@
     <option name="hideEmptyMiddlePackages" value="true" />
     <option name="showLibraryContents" value="true" />
   </component>
-  <component name="PropertiesComponent"><![CDATA[{
-  "keyToString": {
-    "ASKED_ADD_EXTERNAL_FILES": "true",
-    "RunOnceActivity.OpenProjectViewOnStart": "true",
-    "RunOnceActivity.ShowReadmeOnStart": "true",
-    "SHARE_PROJECT_CONFIGURATION_FILES": "true",
-    "WebServerToolWindowFactoryState": "false",
-    "git-widget-placeholder": "PROG-57-excel-file-reader",
-    "ignore.virus.scanning.warn.message": "true",
-    "last_opened_file_path": "/Users/knippenbergl/Programmierprojekt/PP2023/pp-2023/kMeansTestAlgorithm",
-    "list.type.of.created.stylesheet": "SCSS",
-    "node.js.detected.package.eslint": "true",
-    "node.js.detected.package.tslint": "true",
-    "node.js.selected.package.eslint": "(autodetect)",
-    "node.js.selected.package.tslint": "(autodetect)",
-    "nodejs_package_manager_path": "npm",
-    "ts.external.directory.path": "/Applications/IntelliJ IDEA.app/Contents/plugins/javascript-impl/jsLanguageServicesImpl/external",
-    "vue.rearranger.settings.migration": "true"
+  <component name="PropertiesComponent">{
+  &quot;keyToString&quot;: {
+    &quot;ASKED_ADD_EXTERNAL_FILES&quot;: &quot;true&quot;,
+    &quot;RunOnceActivity.OpenProjectViewOnStart&quot;: &quot;true&quot;,
+    &quot;RunOnceActivity.ShowReadmeOnStart&quot;: &quot;true&quot;,
+    &quot;SHARE_PROJECT_CONFIGURATION_FILES&quot;: &quot;true&quot;,
+    &quot;WebServerToolWindowFactoryState&quot;: &quot;false&quot;,
+    &quot;git-widget-placeholder&quot;: &quot;PROG-57-excel-file-reader&quot;,
+    &quot;ignore.virus.scanning.warn.message&quot;: &quot;true&quot;,
+    &quot;last_opened_file_path&quot;: &quot;/Users/knippenbergl/Programmierprojekt/PP2023/pp-2023/kMeansTestAlgorithm&quot;,
+    &quot;list.type.of.created.stylesheet&quot;: &quot;SCSS&quot;,
+    &quot;node.js.detected.package.eslint&quot;: &quot;true&quot;,
+    &quot;node.js.detected.package.tslint&quot;: &quot;true&quot;,
+    &quot;node.js.selected.package.eslint&quot;: &quot;(autodetect)&quot;,
+    &quot;node.js.selected.package.tslint&quot;: &quot;(autodetect)&quot;,
+    &quot;nodejs_package_manager_path&quot;: &quot;npm&quot;,
+    &quot;ts.external.directory.path&quot;: &quot;/Applications/IntelliJ IDEA.app/Contents/plugins/javascript-impl/jsLanguageServicesImpl/external&quot;,
+    &quot;vue.rearranger.settings.migration&quot;: &quot;true&quot;
   }
-}]]></component>
+}</component>
   <component name="RecentsManager">
     <key name="CopyFile.RECENT_KEYS">
       <recent name="$PROJECT_DIR$/pp-2023/kMeansTestAlgorithm" />
@@ -125,7 +124,10 @@
       <workItem from="1695060236105" duration="1040000" />
       <workItem from="1695062691185" duration="7468000" />
       <workItem from="1695230743230" duration="768000" />
-      <workItem from="1696246352566" duration="296000" />
+      <workItem from="1696246352566" duration="1055000" />
+      <workItem from="1696261465480" duration="452000" />
+      <workItem from="1696332543052" duration="11839000" />
+      <workItem from="1696426860569" duration="2587000" />
     </task>
     <task id="LOCAL-00001" summary="add kMeans test algorithm to the repository">
       <option name="closed" value="true" />
@@ -191,7 +193,15 @@
       <option name="project" value="LOCAL" />
       <updated>1695205635634</updated>
     </task>
-    <option name="localTasksCounter" value="9" />
+    <task id="LOCAL-00009" summary="added excel file reader">
+      <option name="closed" value="true" />
+      <created>1696246858695</created>
+      <option name="number" value="00009" />
+      <option name="presentableId" value="LOCAL-00009" />
+      <option name="project" value="LOCAL" />
+      <updated>1696246858695</updated>
+    </task>
+    <option name="localTasksCounter" value="10" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -227,6 +237,7 @@
     <MESSAGE value="run kmeans with react/next.js" />
     <MESSAGE value="impement kmeans" />
     <MESSAGE value="remove kmeans src-folder" />
-    <option name="LAST_COMMIT_MESSAGE" value="remove kmeans src-folder" />
+    <MESSAGE value="added excel file reader" />
+    <option name="LAST_COMMIT_MESSAGE" value="added excel file reader" />
   </component>
 </project>
