+++ pp-2023/Dockerfile
@@ -11,9 +11,6 @@ COPY . .
 # Installieren 
 RUN npm install
 
-# Installieren der xlsx Bibliothek
-RUN npm i xlsx
-
 # Bauen der Next.js-Anwendung
 RUN npm run build
