+++ pp-2023/src/app/kmeans/calculateButton.js
@@ -2,6 +2,7 @@
 import {UseInputKPoints} from './input-k-points';
 import {kMeansAlgorithm} from '../utils/kmeans';
 import {HandleDynamicGeneratedInputFields} from './create-save-manuel-input';
+import ExcelFileUploader from '../components/ExcelFileUploader';
 
 export function HandleCalculateButtonClick() {
 
@@ -14,10 +15,12 @@ export function HandleCalculateButtonClick() {
     const handleClick = () => {
         const kPoints = validateKPoints(numberOfClusters);
         const result = kMeansAlgorithm(dataSet, kPoints);
+
         console.log(result); // Testet Funktion von KMeans
         console.log(inputDataArray); // Testet Funktion der manuellen Eingabe
         console.log(numberOfClusters); // Testet Funktion der K-Eingabe
         return result;
+
     }
 
     /*
